pyproject.toml
[project]
name = "storyforge-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.63.0",
    "google-generativeai>=0.8.6",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.21",
]
